Auth:Create signout action creator, types and actions
