+++ src/redux/action-types/index.js
@@ -1,1 +1,3 @@
-export * from './signin';
+export const SET_CURRENT_USER = 'SET_CURRENT_USER';
+export const SET_CURRENT_USER_ERROR = 'SET_CURRENT_USER_ERROR';
+export const SIGN_OUT = 'SIGN_OUT';
